Changed ranks to longs since bill's views are returning them as long

+++ Product/UpDiddyLib/Dto/SubscriberSourceDto.cs
@@ -7,13 +7,13 @@ namespace UpDiddyLib.Dto
     public class SubscriberSourceDto
     {
 
-        public int? GroupRank { get; set; }
+        public long? GroupRank { get; set; }
         public int SubscriberId { get; set; }
         public Guid? SubscriberGuid { get; set; }
         public string Email { get; set; }
         public string  FirstName { get; set; }
         public string LastName { get; set; }
-        public int? PartnerRank { get; set; }
+        public long? PartnerRank { get; set; }
         public string PartnerName { get; set; }
         public Guid? PartnerGuid { get; set; }
         public string GroupName { get; set; }
